uruti-web/Uruti_Web-updated/src/backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Date/Time
python-dateutil==2.8.2

# File Upload
aiofiles==23.2.1

# Utilities
typing-extensions==4.9.0
firebase-admin==6.5.0

# RAG + multimodal advisory
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
transformers==4.44.2
accelerate==0.34.2
bitsandbytes==0.43.3
pypdf==4.3.1
python-docx==1.1.2
openai-whisper==20240930
